[Bronze II] Title: 숫자, Time: 12 ms, Memory: 2020 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/10093. 숫자/숫자.cc
@@ -5,20 +5,20 @@ int main(){
   ios::sync_with_stdio(0);
   cin.tie(0);
 
-  int a, b;
+  long long a, b;
   cin >> a >> b;
-  int smallNum = min(a, b);
-  int bigNum = max(a, b);
+  long long  smallNum = min(a, b);
+  long long bigNum = max(a, b);
 
   // 두 수 사이의 수 개수
-  int count = 0;
-  for(int i = smallNum + 1; i < bigNum; i++){
+  long long count = 0;
+  for(long long i = smallNum + 1; i < bigNum; i++){
     count++;
   }
   cout << count << "\n";
 
   // 두 수 사이의 수 오름차순
-  for(int i = smallNum + 1; i < bigNum; i++){
+  for(long long i = smallNum + 1; i < bigNum; i++){
     cout << i << " ";
   }
